refactor(fs): move fs logic into cli package
