--- examples/read-file/Cargo.toml
@@ -1,11 +1,0 @@
-[package]
-name = "read-file"
-version = "0.1.0"
-edition = "2021"
-
-[dependencies]
-ytml-lang = {path = "../../packages/ytml-lang", features = ["display", "fs"]}
-
-[[bin]]
-name = "read-file"
-path = "main.rs"
