--- examples/read-file/index.ytml
@@ -1,5 +1,0 @@
-html() {
-    body() {
-        
-    }
-}
