+++ packages/ytml-lang/src/lib.rs
@@ -1,6 +1,3 @@
-#[cfg(feature = "fs")]
-pub mod fs;
 pub mod html;
-pub mod ytml;
 pub mod tokens;
-
+pub mod ytml;
